nix-bindings-sys: add primop tests
I think the main reason we'll use those bindings is to register new
primops without having to patch Nix. Surely the use cases will differ
but this is one of the more "surface level" applications of the
bindings. Let's test registering primops thoroughly so that it does not
come bite us in the ass later.

This is also generously documented to serve as somewhat of a
documentation since I *don't* want to write documentation manually for
everyting, especially since the API will keep changing as I update the
Nix version used to generate the bindings.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: Iddca7a2388679961a41fca68d4514f0c6a6a6964